recipe_creator/requirements.txt
langchain>=1.1.2
langchain-anthropic>=1.2.0
langchain-openai>=1.1.0
langgraph>=1.0.4
langgraph-cli[inmem]>=0.4.7
pydantic>=2.12.0
tavily-python>=0.7.14
youtube-transcript-api>=1.2.0
rich>=14.0.0
python-dotenv>=1.2.0

# Recipe storage
sqlalchemy>=2.0.44
google-cloud-storage>=3.6.0
Pillow>=12.0.0
requests>=2.31.0
langsmith>=0.4.0
loguru>=0.7.0
